Fixed a logical error
When we check the uncompressed size and advance the reader's stream position, we need to leave it there.

The original code doesn't do the safety check, but it does read the uncompressed size and never uses it to decode, only to decide how much to decode; traversing forward before actually decoding.

Signed-off-by: Victor Matia <vmatir@outlook.com>